+++ tests/test_iqm_sampler.py
@@ -16,7 +16,7 @@ import sys
 import uuid
 
 import cirq
-from mockito import ANY, expect, mock, verify, verifyNoUnwantedInteractions, when
+from mockito import ANY, expect, mock, unstub, verify, verifyNoUnwantedInteractions, when
 import numpy as np
 import pytest
 import sympy  # type: ignore
@@ -28,6 +28,7 @@ from iqm.cirq_iqm.iqm_sampler import IQMResult, IQMSampler, ResultMetadata, seri
 from iqm.iqm_client import (
     Circuit,
     CircuitCompilationOptions,
+    CircuitValidationError,
     HeraldingMode,
     Instruction,
     IQMClient,
@@ -99,7 +100,11 @@ def run_request():
 
 @pytest.mark.usefixtures('unstub')
 def test_run_sweep_raises_with_non_physical_names(adonis_sampler, circuit_non_physical):
-    with pytest.raises(ValueError, match='Qubit not on device'):
+    when(adonis_sampler._client).get_quantum_architecture().thenReturn(
+        adonis_sampler._device.metadata.to_architecture()
+    )
+    # Note that validation is done in iqm_client, so this is now an integration test.
+    with pytest.raises(CircuitValidationError, match='Qubit Alice is not allowed as locus for measure'):
         adonis_sampler.run_sweep(circuit_non_physical, None)
 
 
@@ -245,6 +250,10 @@ def test_run_sweep_executes_circuit_with_heralding_mode_zeros(
         'options': CircuitCompilationOptions(heralding_mode=HeraldingMode.ZEROS)
     }
     assert sampler._compiler_options.heralding_mode == HeraldingMode.ZEROS
+    kwargs = create_run_request_default_kwargs | {
+        'options': CircuitCompilationOptions(heralding_mode=HeraldingMode.ZEROS)
+    }
+    assert sampler._compiler_options.heralding_mode == HeraldingMode.ZEROS
     when(client).create_run_request(ANY, **kwargs).thenReturn(run_request)
     when(client).submit_run_request(run_request).thenReturn(job_id)
     when(client).wait_for_results(job_id).thenReturn(run_result)
@@ -329,8 +338,15 @@ def test_run_sweep_abort_job_failed(
 
 @pytest.mark.usefixtures('unstub')
 def test_run_iqm_batch_raises_with_non_physical_names(adonis_sampler, circuit_non_physical):
-    with pytest.raises(ValueError, match='Qubit not on device'):
+    when(adonis_sampler._client).get_quantum_architecture().thenReturn(
+        adonis_sampler._device.metadata.to_architecture()
+    )
+    # Note that validation is done in iqm_client, so this is now an integration test.
+    with pytest.raises(CircuitValidationError, match='Qubit Alice is not allowed as locus for measure'):
         adonis_sampler.run_iqm_batch([circuit_non_physical])
+
+    verifyNoUnwantedInteractions()
+    unstub()
 
 
 @pytest.mark.usefixtures('unstub')